Add cookie based authentication for account login
This change saves the login state of an individual client and alters some pages based on that.
Specifically, if you are logged in, "Login" is changed to "Account" which for now just links to GetAccount. "Placeholder" is replaced with "Logout" which logs you out.

This method of login state saving should persist through closing and opening the site both server-side and client-side.

AccountController had several changes made to accommodate this change.

For now, login state doesn't affect whether you can create events or not, this will likely require a bit of extension to what I have here.